completed test cases for controllers

+++ src/main/java/smartshare/administrationservice/service/BucketAccessRequestService.java
@@ -3,8 +3,8 @@ package smartshare.administrationservice.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import smartshare.administrationservice.constant.StatusConstants;
 import smartshare.administrationservice.dto.AddUserFromUiToBucket;
 import smartshare.administrationservice.dto.BucketAccessRequestFromUi;
 import smartshare.administrationservice.dto.RemoveUserFromBucket;
@@ -16,6 +16,7 @@ import smartshare.administrationservice.repository.BucketRepository;
 import smartshare.administrationservice.repository.UserRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 
 @Slf4j
@@ -44,35 +45,33 @@ public class BucketAccessRequestService {
         this.addUserFromUiToBucketMapper = addUserFromUiToBucketMapper;
     }
 
-    public Status createBucketAccessRequest(BucketAccessRequestFromUi bucketAccessRequestFromUi) {
+
+    public Boolean createBucketAccessRequest(BucketAccessRequestFromUi bucketAccessRequestFromUi) {
         log.info( "Inside createBucketAccessRequest" );
         try {
             BucketAccessRequest newBucketAccessRequest = bucketAccessRequestMapper.map( bucketAccessRequestFromUi );
             BucketAccessRequest createdBucketAccessRequest = bucketAccessRequestRepository.saveAndFlush( newBucketAccessRequest );
             System.out.println( "createdBucketAccessRequest------------->" + createdBucketAccessRequest );
-            statusOfOperation.setMessage( StatusConstants.SUCCESS.toString() );
+            return Boolean.TRUE;
         } catch (Exception e) {
             log.error( "Exception while creating bucket Access Request" + e.getMessage() );
-            statusOfOperation.setMessage( StatusConstants.FAILED.toString() );
         }
-        return statusOfOperation;
+        return Boolean.FALSE;
     }
 
     private Bucket addUserToBucket(Bucket bucketInWhichUserListToBeUpdated, UserBucketMapping userBucketMapping) {
         log.info( "Inside addUserToBucket" );
         bucketInWhichUserListToBeUpdated.addUser( userBucketMapping );
         return bucketRepository.saveAndFlush( bucketInWhichUserListToBeUpdated );
-
     }
 
-    private UserBucketMapping isUserAddedInBucket(Bucket bucket, User user) {
+    private Optional<UserBucketMapping> isUserAddedInBucket(Bucket bucket, User user) {
         log.info( "Inside isUserAddedInBucket" );
 
         List<UserBucketMapping> usersInTheBucket = bucket.getAccessingUsers();
         return usersInTheBucket.stream()
                 .filter( userBucketMapping -> userBucketMapping.getUser().equals( user ) )
-                .findAny()
-                .orElse( null );
+                .findAny();
 
     }
 
@@ -83,11 +82,12 @@ public class BucketAccessRequestService {
     }
 
     private void approveWriteAccessRequestHandler(BucketAccessRequest bucketAccessRequest) {
-        UserBucketMapping userExistsInBucket = isUserAddedInBucket( bucketAccessRequest.getBucket(), bucketAccessRequest.getUser() );
-        if (userExistsInBucket != null) {
-            addUserToBucket( bucketAccessRequest.getBucket(), giveWritePermissionToTheUserInTheBucket( userExistsInBucket ) );
+        Optional<UserBucketMapping> userExistsInBucket = isUserAddedInBucket( bucketAccessRequest.getBucket(), bucketAccessRequest.getUser() );
+        if (userExistsInBucket.isPresent()) {
+            addUserToBucket( bucketAccessRequest.getBucket(), giveWritePermissionToTheUserInTheBucket( userExistsInBucket.get() ) );
         } else {
-            UserBucketMapping newUserBucketMappingWithWriteAccess = new UserBucketMapping( bucketAccessRequest.getUser(), bucketAccessRequest.getBucket(), new BucketAccess( Boolean.TRUE, Boolean.TRUE ) );
+            UserBucketMapping newUserBucketMappingWithWriteAccess = new UserBucketMapping(
+                    bucketAccessRequest.getUser(), bucketAccessRequest.getBucket(), new BucketAccess( Boolean.TRUE, Boolean.TRUE ) );
             addUserToBucket( bucketAccessRequest.getBucket(), newUserBucketMappingWithWriteAccess );
         }
     }
@@ -95,70 +95,74 @@ public class BucketAccessRequestService {
     private void createOrUpdateAccessRecordsBasedOnTheRequest(BucketAccessRequest bucketAccessRequest) {
         log.info( "Inside createOrUpdateAccessRecordsBasedOnTheRequest" );
         if (bucketAccessRequest.getAccess().getRead())
-            addUserToBucket( bucketAccessRequest.getBucket(), new UserBucketMapping( bucketAccessRequest.getUser(), bucketAccessRequest.getBucket(), bucketAccessRequest.getAccess() ) );
+            addUserToBucket(
+                    bucketAccessRequest.getBucket(),
+                    new UserBucketMapping(
+                            bucketAccessRequest.getUser(),
+                            bucketAccessRequest.getBucket(),
+                            bucketAccessRequest.getAccess() ) );
         if (bucketAccessRequest.getAccess().getWrite()) approveWriteAccessRequestHandler( bucketAccessRequest );
 
     }
 
-    public Status approveBucketAccessRequest(BucketAccessRequest bucketAccessRequest) {
+    public Boolean approveBucketAccessRequest(BucketAccessRequest bucketAccessRequest) {
         log.info( "Inside approveBucketAccessRequest" );
         try {
             bucketAccessRequest.approve();
             createOrUpdateAccessRecordsBasedOnTheRequest( bucketAccessRequest );
             bucketAccessRequestRepository.saveAndFlush( bucketAccessRequest );
-            statusOfOperation.setMessage( StatusConstants.SUCCESS.toString() );
+            return Boolean.TRUE;
         } catch (Exception e) {
             log.error( "Exception while approving the Bucket Access Request" + e.getMessage() );
-            statusOfOperation.setMessage( StatusConstants.FAILED.toString() );
         }
-        return statusOfOperation;
+        return Boolean.FALSE;
     }
 
-    public Status rejectBucketAccessRequest(BucketAccessRequest bucketAccessRequest) {
+    public Boolean rejectBucketAccessRequest(BucketAccessRequest bucketAccessRequest) {
         log.info( "Inside approveBucketAccessRequest" );
         try {
             bucketAccessRequest.reject();
             bucketAccessRequestRepository.saveAndFlush( bucketAccessRequest );
-            statusOfOperation.setMessage( StatusConstants.SUCCESS.toString() );
+            return Boolean.TRUE;
         } catch (Exception e) {
             log.error( "Exception while rejecting the Bucket Access Request" + e.getMessage() );
-            statusOfOperation.setMessage( StatusConstants.FAILED.toString() );
         }
-        return statusOfOperation;
+        return Boolean.FALSE;
     }
 
 
-    public Status addUserToBucketByBucketAdmin(AddUserFromUiToBucket addUserFromUiToBucket) {
+    public Boolean addUserToBucketByBucketAdmin(AddUserFromUiToBucket addUserFromUiToBucket) {
         log.info( "Inside addUserToBucket" );
         try {
             AddUserFromUiToBucketMapper mappedResult = addUserFromUiToBucketMapper.map( addUserFromUiToBucket );
             addUserToBucket( mappedResult.getBucketToWhichUserIsAdded(), mappedResult.getNewUserMapping() );
-            statusOfOperation.setMessage( StatusConstants.SUCCESS.toString() );
+            return Boolean.TRUE;
         } catch (Exception e) {
             log.error( "Exception while adding user to the Bucket by admin" + e.getMessage() );
-            statusOfOperation.setMessage( StatusConstants.FAILED.toString() );
         }
-        return statusOfOperation;
+        return Boolean.FALSE;
     }
 
     public Status removeUserFromBucketByBucketAdmin(RemoveUserFromBucket removeUserFromBucket) {
         log.info( "Inside removeUserToBucketByBucketAdmin" );
         try {
-            Bucket bucketFromWhichUserIsRemoved = bucketRepository.findByName( removeUserFromBucket.getBucketName() );
-            User user = userRepository.findByUserName( removeUserFromBucket.getUserName() );
-            UserBucketMapping userBucketMapping = isUserAddedInBucket( bucketFromWhichUserIsRemoved, user );
-            if (bucketFromWhichUserIsRemoved.getObjects() == null) {
-                bucketFromWhichUserIsRemoved.removeUser( userBucketMapping );
-                bucketRepository.saveAndFlush( bucketFromWhichUserIsRemoved );// not sure whether it works or not
-                statusOfOperation.setMessage( StatusConstants.SUCCESS.toString() );
-            } else {
-                statusOfOperation.setMessage( StatusConstants.FAILED.toString() );
+            Optional<Bucket> bucketFromWhichUserIsRemoved = Optional.ofNullable( bucketRepository.findByName( removeUserFromBucket.getBucketName() ) );
+            Optional<User> user = Optional.ofNullable( userRepository.findByUserName( removeUserFromBucket.getUserName() ) );
+
+            if (bucketFromWhichUserIsRemoved.isPresent() && user.isPresent()) {
+                Optional<UserBucketMapping> doesUserBucketMappingExists = isUserAddedInBucket( bucketFromWhichUserIsRemoved.get(), user.get() );
+                if (doesUserBucketMappingExists.isPresent() && bucketFromWhichUserIsRemoved.get().getObjects().isEmpty()) {
+                    bucketFromWhichUserIsRemoved.get().removeUser( doesUserBucketMappingExists.get() );
+                    bucketRepository.saveAndFlush( bucketFromWhichUserIsRemoved.get() );// not sure whether it works or not
+                    statusOfOperation.setValue( Boolean.TRUE );
+                } else {
+                    statusOfOperation.setValue( Boolean.FALSE );
                 statusOfOperation.setReasonForFailure( "The user has Objects which has to be deleted before removing the user" );
             }
-
+            }
         } catch (Exception e) {
             log.error( "Exception while removing user from the Bucket by admin" + e.getMessage() );
-            statusOfOperation.setMessage( StatusConstants.FAILED.toString() );
+            statusOfOperation.setValue( Boolean.FALSE );
         }
         return statusOfOperation;
     }
@@ -173,11 +177,18 @@ public class BucketAccessRequestService {
 
     public Status deleteBucketAccessRequest(BucketAccessRequest bucketAccessRequestTobeDeleted) {
         try {
-            bucketAccessRequestRepository.delete( bucketAccessRequestTobeDeleted );
-            statusOfOperation.setMessage( StatusConstants.SUCCESS.toString() );
+            Optional<BucketAccessRequest> doesBucketAccessRequestWhichWillBeDeletedExists = bucketAccessRequestRepository.findById( bucketAccessRequestTobeDeleted.getId() );
+            if (doesBucketAccessRequestWhichWillBeDeletedExists.isPresent()) {
+                bucketAccessRequestRepository.delete( bucketAccessRequestTobeDeleted );
+                statusOfOperation.setValue( Boolean.TRUE );
+            } else {
+                statusOfOperation.setValue( Boolean.FALSE );
+                statusOfOperation.setReasonForFailure( HttpStatus.NOT_FOUND.getReasonPhrase() );
+            }
         } catch (Exception e) {
             log.error( "Exception while deleting the bucket access request by admin" + e.getMessage() );
-            statusOfOperation.setMessage( StatusConstants.FAILED.toString() );
+            statusOfOperation.setValue( Boolean.FALSE );
+            statusOfOperation.setReasonForFailure( HttpStatus.INTERNAL_SERVER_ERROR.getReasonPhrase() );
         }
         return statusOfOperation;
